refactor: moves `lambda.ts` from `data/services` to `/data`
refactor: renames `Dropzone` component to `ImageUpload`

refactor: removes unnecessary `PostsForm` wrapper component

changes after reviewing the code
